+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Updated FAG index tooltip styles (mobile / tablet)
 - Updated 'Market' section to store row and page number state when returning from other page
 - Enhanced user navigation with scroll position retention for coin list
+- Introduced CoinGecko context and refactored Home & Market components to reduce number of API calls made
 
 # v1.0.0 - 1st October 2023